% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highway_authorities}
\alias{highway_authorities}
\alias{highway_authorities_mod}
\title{Highway authorities}
\format{An sf object}
\description{
These are the highway authorities in the UK.
`highway_authorities_mod` is a modified version for plotting.
}
\examples{
\dontrun{
highway_authorities = sf::read_sf("data/boundaries/ha.gpkg")
highway_authorities = rmapshaper::ms_simplify(highway_authorities)
# highway_authorities = sf::st_simplify(highway_authorities, 0.001)
pryr::object_size(highway_authorities)
}
sf:::plot.sf(highway_authorities)
sf:::plot.sf(highway_authorities_mod)
}
\keyword{datasets}
