% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msoa_ashford}
\alias{msoa_ashford}
\title{msoa boundary data for ashford}
\format{An sf object}
\description{
A sample of UK administrative boundaries for educational purposes
for Ashford, UK.
}
\examples{
\dontrun{
msoa_uk = sf::read_sf("data/boundaries/msoa.gpkg")
pryr::object_size(msoa_uk) # 66 MB
msoa_uk_cent = sf::st_centroid(msoa_uk)
msoa_ashford_cent = msoa_uk_cent[ashford_bng, ]
msoa_ashford = msoa_uk[msoa_uk$id \%in\% msoa_ashford$id, ]
}
plot(msoa_ashford$geom)
plot(htdd_ashford$i__location_point, add = TRUE)
}
\keyword{datasets}
