% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road_stats}
\alias{road_stats}
\title{Road statistics for HAs}
\format{An sf object}
\description{
Summary statistics on roads in highway authorities.
}
\examples{
\dontrun{
road_stats = readRDS("data/HA_RoadStats.Rds")
usethis::use_data(road_stats)
}
summary(road_stats$name \%in\% highway_authorities$name)
road_stats
}
\keyword{datasets}
