% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_import_elgin.R
\name{rw_clean_points}
\alias{rw_clean_points}
\title{Clean Spatial Part of the HDDT Data}
\usage{
rw_clean_points(df)
}
\arguments{
\item{df}{a data frame to clean}
}
\description{
From https://downloads.srwr.scot/export
}
