% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_import_elgin.R
\name{rw_import_elgin_batch}
\alias{rw_import_elgin_batch}
\title{Bulk Import Elgin road works data}
\usage{
rw_import_elgin_batch(folder_path, method = c("ttvdd", "ed",
  "restrictions", "htdd"), pattern, ...)
}
\arguments{
\item{folder_path}{path to folder containing csv file or compressed csv file}

\item{method}{name of import algorithum to use}

\item{pattern}{character string containing a regular expression. This is used as a filter rule for file names. For example pattern = "test" whold only import files with test in the file name. Passed to base::grepl}

\item{...}{additonal variaibles passed to grepl}
}
\description{
This function imports data that has already been downloaded from
http://downloads.roadworks.org/
}
\details{
A useful pattern is 'pattern = '^(?!.*log).*ttvdd'
Which will return all files names with 'ttvdd' but exclude all file names with 'log'
You must also pass the perl = TRUE parameter
}
\examples{
\dontrun{
d = rw_import_elgin_batch("C:/data","rw_import_elgin_ttvdd", '^(?!.*log).*ttvdd', perl = TRUE)
}
}
