% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_OS.R
\name{rw_import_OSMM}
\alias{rw_import_OSMM}
\title{Function TO Read in OS MasterMap map GML}
\usage{
rw_import_OSMM(folder)
}
\arguments{
\item{folder}{character, path to folder of GML.gz file of OSMM
Return a named list of SF data frames.
N.B. non-spatial GML files not yet supported}
}
\description{
This function helps read-in Ordnance Survey data
}
