% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_import_scotland.R
\name{rw_import_scot}
\alias{rw_import_scot}
\title{Import Scottish road works data}
\usage{
rw_import_scot(zip_file, dest_dir = tempdir())
}
\description{
This function imports data that has already been downloaded from
https://downloads.srwr.scot/export
}
\examples{
\dontrun{
zip_file = "data/01.zip"
d = rw_import_scot(zip_file)
ncol(d) # 74 columns
}
}
