% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_map_interactive.R
\name{rw_map_interactive}
\alias{rw_map_interactive}
\title{Create an interactive map of htdd data}
\usage{
rw_map_interactive(x, vars = names(x), ...)
}
\arguments{
\item{x}{A dataset containing htdd data object created with [rw_import_elgin_htdd()]}

\item{vars}{A text string of variables to be show, one facet for each}

\item{...}{Arguments passed to `leaflet()`}
}
\description{
This function creates a simple static map of roadworks data, for one or more variables.
}
\examples{
rw_map_interactive(htdd_ashford[1:100, ])
rw_map_interactive(htdd_ashford[1:100, ], vars = names(htdd_ashford)[1:3])
}
