% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_gis.R
\name{rw_points_to_region}
\alias{rw_points_to_region}
\title{Match Points to Regions}
\usage{
rw_points_to_region(points, regions, ncores = 1)
}
\arguments{
\item{points}{simple features column of points}

\item{regions}{simple features data frame of polygons}

\item{ncores}{number of core to run (default = 1)}
}
\description{
Match Points to Regions
}
\examples{
\dontrun{
zip_file = "data/01.zip"
d = rw_import_scot(zip_file)
ncol(d) # 74 columns
}
}
