% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_spatial.R
\name{rw_spatial}
\alias{rw_spatial}
\title{Convert an data frame containing roadworks data into a spatial data object}
\usage{
rw_spatial(x)
}
\arguments{
\item{x}{param}
}
\description{
This function will take a data frame and make it spatial
}
