% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_spec.R
\name{rw_spec}
\alias{rw_spec}
\title{View specification documents associated with data provided by Elgin}
\usage{
rw_spec(spec_type = c("htdd", "enquires", "restrictions", "ttvdd"))
}
\arguments{
\item{spec_type}{the type of specification document, one of:
htdd, enquires, restrictions or ttvdd.}
}
\description{
View specification documents associated with data provided by Elgin
}
\examples{
\dontrun{
rw_spec()
rw_spec("restrictions")
}
}
