% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_urls}
\alias{spec_urls}
\title{URLs pointing to roadworks data spec from Elgin}
\format{Character vector}
\description{
URLs pointing to roadworks data spec from Elgin
}
\examples{
\dontrun{
f = list.files(pattern = "pdf")
piggyback::pb_upload(f)
u = piggyback::pb_download_url()
spec_urls = u[grepl(pattern = "Elgin", x = u)]
road_stats = readRDS("data/HA_RoadStats.Rds")
usethis::use_data(road_stats)
}
spec_urls
spec_urls[grepl("HTDD", spec_urls)]
}
\keyword{datasets}
